Do not try to unlink tcp-connection as a file

+++ gck-rpc-dispatch.c
@@ -2389,7 +2389,8 @@ void gck_rpc_layer_uninitialize(void)
 	pkcs11_socket = -1;
 
 	/* Delete our unix socket */
-	if (pkcs11_socket_path[0])
+	if (pkcs11_socket_path[0] &&
+		strncmp(pkcs11_socket_path, "tcp://", strlen("tcp://")) != 0)
 		unlink(pkcs11_socket_path);
 	pkcs11_socket_path[0] = 0;
